WIP: Refactored but SharedTreeView has public constructor and API extractor doesn't like that.
