+++ experimental/dds/tree2/src/index.ts
@@ -258,9 +258,9 @@ export {
 	identifierKeySymbol,
 	ISharedTree,
 	ISharedTreeView,
-	ISharedTreeFork,
 	runSynchronous,
 	SharedTreeFactory,
+	SharedTreeView,
 	ViewEvents,
 	SchematizeConfiguration,
 } from "./shared-tree";
